Update user delete system

+++ Chrono-PM-API/Services/UserService.cs
@@ -63,8 +63,61 @@ public class UserService : IUserService
         {
             throw new ArgumentException($"User not found with id: {id}");
         }
+        
+        if (user.UserName == "DeletedUser")
+        {
+            await _userManager.DeleteAsync(user);
+        }
+        
+        if (user.IsActive)
+        {
+            user.IsActive = false;
+            var updateResult = await _userManager.UpdateAsync(user);
+            if (!updateResult.Succeeded)
+            {
+                throw new Exception($"Failed to update user: {updateResult.Errors}");
+            }
+        }
+        else
+        {
+            user.UserName = "DeletedUser";
+            user.Firstname = "Deleted";
+            user.Lastname = "User";
+            user.Email = "no email";
+            user.PhoneNumber = "no phone";
 
-        await _userManager.DeleteAsync(user);
+            var updateResult = await _userManager.UpdateAsync(user);
+            if (!updateResult.Succeeded)
+            {
+                throw new Exception($"Failed to update user: {updateResult.Errors}");
+            }
+
+            return true;
+        }
         return true;
+    }
+    
+    public async Task<bool> ChangeUserActivationStatusAsync(string id, bool status)
+    {
+        var user = await _userManager.FindByIdAsync(id);
+        if (user == null)
+        {
+            throw new ArgumentException($"User not found with id: {id}");
+        }
+
+        if (user.IsActive == status)
+        {
+            throw new ArgumentException($"User status already {(user.IsActive ? "Active" : "Inactive")}");
+        }
+        
+        user.IsActive = status;
+
+        var result = await _userManager.UpdateAsync(user);
+        if (!result.Succeeded)
+        {
+            throw new Exception($"Failed to update user: {result.Errors}");
+        }
+        
+        return user.IsActive;
     }
 }
